refactor: common package 삭제 및 도메인 비즈니스 예외 도메인 계층으로 변경
