+++ src/main/java/roomescape/application/reservation/ReservationService.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Service;
 import roomescape.application.reservation.dto.CreateReservationParam;
 import roomescape.application.reservation.dto.ReservationResult;
 import roomescape.application.reservation.dto.ReservationSearchParam;
-import roomescape.common.exception.BusinessRuleViolationException;
-import roomescape.common.exception.NotFoundEntityException;
+import roomescape.application.support.exception.NotFoundEntityException;
+import roomescape.domain.BusinessRuleViolationException;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRepository;
 import roomescape.domain.reservation.Reservation;
