+++ src/main/java/roomescape/application/reservation/ReservationTimeService.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Service;
 import roomescape.application.reservation.dto.AvailableReservationTimeResult;
 import roomescape.application.reservation.dto.CreateReservationTimeParam;
 import roomescape.application.reservation.dto.ReservationTimeResult;
-import roomescape.common.exception.BusinessRuleViolationException;
-import roomescape.common.exception.NotFoundEntityException;
+import roomescape.application.support.exception.NotFoundEntityException;
+import roomescape.domain.BusinessRuleViolationException;
 import roomescape.domain.reservation.DailyThemeReservations;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservation.ReservationRepository;
