+++ src/main/java/roomescape/application/reservation/ThemeService.java
@@ -6,8 +6,8 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.application.reservation.dto.CreateThemeParam;
 import roomescape.application.reservation.dto.ThemeResult;
-import roomescape.common.exception.BusinessRuleViolationException;
-import roomescape.common.exception.NotFoundEntityException;
+import roomescape.application.support.exception.NotFoundEntityException;
+import roomescape.domain.BusinessRuleViolationException;
 import roomescape.domain.reservation.ReservationRepository;
 import roomescape.domain.reservation.Theme;
 import roomescape.domain.reservation.ThemeRepository;
