+++ src/main/java/roomescape/application/support/exception/ForbiddenException.java
@@ -1,4 +1,4 @@
-package roomescape.common.exception;
+package roomescape.application.support.exception;
 
 public class ForbiddenException extends CoreException {
     public ForbiddenException(String message) {
