+++ src/test/java/roomescape/application/ReservationServiceTest.java
@@ -12,8 +12,8 @@ import roomescape.application.reservation.dto.CreateReservationParam;
 import roomescape.application.reservation.dto.ReservationResult;
 import roomescape.application.reservation.dto.ReservationTimeResult;
 import roomescape.application.reservation.dto.ThemeResult;
-import roomescape.common.exception.BusinessRuleViolationException;
-import roomescape.common.exception.NotFoundEntityException;
+import roomescape.application.support.exception.NotFoundEntityException;
+import roomescape.domain.BusinessRuleViolationException;
 import roomescape.domain.member.Email;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.Role;
