+++ src/test/java/roomescape/application/ReservationTimeServiceTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.Test;
 import roomescape.application.reservation.dto.AvailableReservationTimeResult;
 import roomescape.application.reservation.dto.CreateReservationTimeParam;
 import roomescape.application.reservation.dto.ReservationTimeResult;
-import roomescape.common.exception.BusinessRuleViolationException;
-import roomescape.common.exception.NotFoundEntityException;
+import roomescape.application.support.exception.NotFoundEntityException;
+import roomescape.domain.BusinessRuleViolationException;
 import roomescape.domain.member.Email;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.Role;
