+++ src/test/java/roomescape/domain/DailyThemeReservationsTest.java
@@ -8,7 +8,6 @@ import java.time.LocalTime;
 import java.util.List;
 import java.util.Set;
 import org.junit.jupiter.api.Test;
-import roomescape.common.exception.BusinessRuleViolationException;
 import roomescape.domain.member.Email;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.Role;
