+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -6,7 +6,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.time.LocalTime;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import roomescape.common.exception.BusinessRuleViolationException;
 import roomescape.domain.reservation.ReservationTime;
 
 class ReservationTimeTest {
